Move stuff around to keep impl closer to their definitions
This is all very silly, but more readable.  So much confusion.
